pyproject.toml
[tool.poetry]
name = "ms-user-candidate"
version = "0.1.0"
description = ""
authors = ["Real Huseyn <realhuseynli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
email-validator = "^1.1.2"
gunicorn = "^20.0.4"
uvloop = "^0.16"
httptools = "^0.3.0"
pymongo = "^4.3.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
